Communication AGM -> ACE
