+++ bwplus_communications/$PREFIX$
@@ -1,0 +1,1 @@
+y\BWplus\addons\communications
