--- bwplus_communications/fnc/fn_canConnectSatcom.sqf
@@ -1,10 +1,0 @@
-private ["_result", "_frequency"];
-
-_result = false;
-_frequency = (call TFAR_fnc_ActiveLRRadio) call TFAR_fnc_getLRFrequency;
-
-if (call TFAR_fnc_isTeamSpeakPluginEnabled && {!isnil "_frequency"}) then {
-	_result = true;
-};
-
-_result
