+++ bwplus_communications/functions/fnc_isConnected.sqf
@@ -1,0 +1,2 @@
+#include "script_component.hpp"
+/// Obsolent?
